Abstract apollo client into its own config file
