+++ src/providers/ApolloProvider.tsx
@@ -1,34 +1,11 @@
-import {
-    ApolloClient,
-    ApolloProvider as Provider,
-    createHttpLink,
-    InMemoryCache,
-} from '@apollo/client';
-import { setContext } from '@apollo/client/link/context';
+import { ApolloProvider as Provider } from '@apollo/client';
 
-const httpLink = createHttpLink({
-    uri: 'http://localhost:4000',
-});
-
-const authLink = setContext((_, { headers }) => {
-    const token = localStorage.getItem('token');
-    return {
-        headers: {
-            ...headers,
-            authorization: token ? `Bearer ${JSON.parse(token)}` : '',
-        },
-    };
-});
-
-const client = new ApolloClient({
-    link: authLink.concat(httpLink),
-    cache: new InMemoryCache(),
-});
+import { apolloClient } from '@/lib/apolloClient';
 
 type ApolloProviderProps = {
     children: React.ReactNode;
 };
 
 export const ApolloProvider = ({ children }: ApolloProviderProps) => {
-    return <Provider client={client}>{children}</Provider>;
+    return <Provider client={apolloClient}>{children}</Provider>;
 };
